backend/app/requirements.txt
# מחזיק את כל הספריות שהפרויקט צריך בשביל לרוץ

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
bcrypt
python-jose[cryptography]
apscheduler
pytz
pytest
httpx
pytest-asyncio
google-auth
google-auth-oauthlib
google-api-python-client
